# api/calc.swagger.yaml
openapi: 3.0.1
info:
  title: Yandex Go 2025 Calc service
  version: version not set
  description: >
    API для конкурентного вычисления ООООЧЕНЬ сложных математических операций(калькулятор)
...
servers:
- url: http://localhost:9090/api/v1
paths:
...
  /expressions/{id}:
    get:
      tags:
        - Expressions
      summary: Получение выражения по ID
      description: Возвращает выражение по ID
      operationId: getExpression
      parameters:
        - in: path
          name: id
          required: true
          description: ID выражения
          schema:
            type: number
            format: int64
      responses:
        "200":
          description: Список выражений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/createExpressionResp'
        "404":
          description: Выражений не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Внутренняя ошибка сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      description: Ответ с описанием ошибки
      properties:
        status:
          type: string
          description: Статус ответа
          example: "error"
        message:
          type: string
          description: Информация о ошибке
          example: "error message"
      required:
        - status
        - message
...
    createExpressionResp:
      type: object
      properties:
        id:
          type: integer
          description: ID созданного выражения
          format: int64
          example: 0
      required:
        - id
...